+++ NEWS.md
@@ -1,3 +1,5 @@
+# vapour (development version)
+
 # vapour 0.4.0
 
 ## BREAKING CHANGES